Refactor agentic packages from pkg into system/agentic
