+++ server/app/boot_levels.go
@@ -45,10 +45,10 @@ import (
 	sysService "github.com/cortezaproject/corteza/server/system/service"
 	sysEvent "github.com/cortezaproject/corteza/server/system/service/event"
 	"github.com/cortezaproject/corteza/server/system/types"
-	"github.com/cortezaproject/corteza/server/pkg/agentic/observability"
+	"github.com/cortezaproject/corteza/server/system/agentic/observability"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
-	mcpkg "github.com/cortezaproject/corteza/server/system/mcp"
+	mcpkg "github.com/cortezaproject/corteza/server/system/agentic/mcp"
 	"github.com/lestrrat-go/jwx/jwt"
 	"go.uber.org/zap"
 	gomail "gopkg.in/mail.v2"
