+++ server/system/agentic/mcp/record_handler.go
