+++ server/system/agentic/registry/registry.go
