+++ server/system/agentic/registry/registry_test.go
