+++ server/system/service/service.go
@@ -11,8 +11,8 @@ import (
 	automationService "github.com/cortezaproject/corteza/server/automation/service"
 	discoveryService "github.com/cortezaproject/corteza/server/discovery/service"
 	"github.com/cortezaproject/corteza/server/pkg/actionlog"
-	"github.com/cortezaproject/corteza/server/pkg/agentic/observability"
-	agenticRuntime "github.com/cortezaproject/corteza/server/pkg/agentic/runtime"
+	"github.com/cortezaproject/corteza/server/system/agentic/observability"
+	agenticRuntime "github.com/cortezaproject/corteza/server/system/agentic/runtime"
 	"github.com/cortezaproject/corteza/server/pkg/dal"
 	"github.com/cortezaproject/corteza/server/pkg/eventbus"
 	"github.com/cortezaproject/corteza/server/pkg/healthcheck"
